requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
jinja2
python-multipart
pymupdf
sentence-transformers
faiss-cpu
tqdm
torch
transformers
accelerate
